unbundle svelte generator
