+++ packages/core/src/generators/svelte/index.ts
@@ -1,0 +1,2 @@
+export { componentToSvelte } from './svelte';
+export { ToSvelteOptions } from './types';
