+++ packages/core/src/generators/svelte/types.ts
@@ -1,0 +1,5 @@
+import { BaseTranspilerOptions } from '../../types/transpiler';
+
+export interface ToSvelteOptions extends BaseTranspilerOptions {
+  stateType?: 'proxies' | 'variables';
+}
